Adds base elements for localization
This commits adds the necessary elements for the localization of Ajour.
This also provides a first draft of the french localization, translated
by myself.

This is done based on the GNU gettext tools ported to rust and the
translations are embedded into the binary with i18n-embed. At the code
level, all there is to do is to add the `tr!()` macro to the text that
needs to be translated, and running `cargo i18n` will pick those strings
up and set them up in a `.po` file ready to be translated.

This changes the build process a little bit as it requires external
tools to create the localization files and package them when building.

This will require the following on the builder's machine :
- `cargo install xtr`
- `cargo install cargo-i18n`

Signed-off-by: Jérémie Carpentier-Roy <jeremie.carpentier.roy@gmail.com>